Share Z.ai base url handling and fix flash context fallback
